requirements.txt
fastapi
pydantic
uvicorn
numexpr
hera
pyyaml
openai
python-dotenv